Extract hook in own file
